fix(asf): address review pass 2 findings (NM2/NS1-3/NI1)
N-M1 (branch sync with origin/main) landed in the prior merge
commit; this commit covers the remaining items:

N-M2: adds the unit test for the M1 ValueError branch in _fetch —
a RemoteProduct with `fileName: None` now exercises the
"resolvable fileName" error path, lifting backend.py to 100% line +
branch coverage; the test also asserts the auth wrapper is not
configured (fail-fast before paying the EDL login cost).

N-S1: docs/reference/asf/usage.md was still telling users to pass
`temporal_baseline=(0.0, 60.0)`; floats would route through
_stack_opts as the string "0.0,60.0", which the SDK rejects.
Updated to `(0, 60)` to match the narrowed `tuple[int, int]` type.

N-S2: the anonymous e2e search test now `pytest.skip(...)`s when
the rift-zone window returns zero products. ASF availability is
upstream-controlled; a quiet window is not a backend regression.

N-S3: `_helpers.apply_baseline_windows` `temporal_baseline`
parameter narrowed to `tuple[int, int] | None` to match the public
API change from pass 1.

N-I1: drop the unused `unittest.mock.MagicMock` import in
tests/asf/conftest.py.

+++ src/earthlens/asf/_helpers.py
@@ -89,7 +89,7 @@ def _in_window(value: float | int | None, window: tuple[float, float] | None) ->
 def apply_baseline_windows(
     products: list[Any],
     perpendicular_baseline: tuple[float, float] | None,
-    temporal_baseline: tuple[float, float] | None,
+    temporal_baseline: tuple[int, int] | None,
 ) -> list[Any]:
     """Drop stacked products outside the requested baseline windows.
